refactor: Application을 멀티모듈로 분리한다.
