+++ application/src/test/java/kitchenpos/ApplicationTest.java
