+++ application/src/test/java/kitchenpos/application/MenuGroupServiceIntegrateTest.java
@@ -1,9 +1,11 @@
 package kitchenpos.application;
 
-import kitchenpos.application.test.ServiceIntegrateTest;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.repository.MenuGroupRepository;
-import kitchenpos.dto.response.MenuGroupResponse;
+import kitchenpos.execute.ServiceIntegrateTest;
+import kitchenpos.fixture.MenuGroupFixture;
+import kitchenpos.menu.application.MenuGroupService;
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.domain.repository.MenuGroupRepository;
+import kitchenpos.menu.dto.response.MenuGroupResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
@@ -11,7 +13,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 import java.util.List;
 
-import static kitchenpos.domain.fixture.MenuGroupFixture.인기_메뉴_생성;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
@@ -29,7 +30,7 @@ class MenuGroupServiceIntegrateTest extends ServiceIntegrateTest {
         @Test
         void 모든_메뉴_그룹을_생성한다() {
             // given
-            MenuGroup 인기_메뉴 = 인기_메뉴_생성();
+            MenuGroup 인기_메뉴 = MenuGroupFixture.인기_메뉴_생성();
 
             // when, then
             assertDoesNotThrow(() -> menuGroupService.create(인기_메뉴.getName()));
@@ -42,7 +43,7 @@ class MenuGroupServiceIntegrateTest extends ServiceIntegrateTest {
 
         @BeforeEach
         void setUp() {
-            MenuGroup 인기_메뉴 = 인기_메뉴_생성();
+            MenuGroup 인기_메뉴 = MenuGroupFixture.인기_메뉴_생성();
             menuGroupRepository.save(인기_메뉴);
         }
