+++ application/src/test/java/kitchenpos/execute/ServiceIntegrateTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.test;
+package kitchenpos.execute;
 
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.springframework.boot.test.context.SpringBootTest;
