--- product/build.gradle
@@ -1,19 +1,0 @@
-plugins {
-    id 'java'
-}
-
-group 'camp.nextstep.edu'
-version '0.0.1-SNAPSHOT'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
-}
-
-test {
-    useJUnitPlatform()
-}
